+++ assets/js/app.js
@@ -7,10 +7,20 @@ import {LiveSocket} from "phoenix_live_view"
 
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content")
 let liveSocket = new LiveSocket("/live", Socket, {params: {_csrf_token: csrfToken}})
+let topBarScheduled = undefined
 
 topbar.config({barColors: {0: "#29d"}, shadowColor: "rgba(0, 0, 0, .3)"})
-window.addEventListener("phx:page-loading-start", info => topbar.show())
-window.addEventListener("phx:page-loading-stop", info => topbar.hide())
+
+window.addEventListener("phx:page-loading-start", (info) => {
+  if(!topBarScheduled) {
+    topBarScheduled = setTimeout(() => topbar.show(), 1)
+  }
+})
+window.addEventListener("phx:page-loading-stop", (info) => {
+  clearTimeout(topBarScheduled)
+  topBarScheduled = undefined
+  topbar.hide()
+})
 
 liveSocket.connect()